+++ app/domain/data_objects/data_object_mapping.rb
@@ -20,7 +20,8 @@ module DataObjects
       diff_data[:resources].each do |obj|
         resource_type = obj[:resource_id].split(':')[1]
         next unless role_types.include?(resource_type)
-        items[obj[:resource_id]] = obj
+
+        items[obj[:resource_id]] = obj.clone
       end
 
       # Now map attributes, such as annotations, role_memberships, credentials,
@@ -68,7 +69,8 @@ module DataObjects
       diff_data[:resources].each do |obj|
         resource_type = obj[:resource_id].split(':')[1]
         next unless resource_types.include?(resource_type)
-        items[obj[:resource_id]] = obj
+
+        items[obj[:resource_id]] = obj.clone
       end
 
       # [schema_attribute, primitive_attribute, primitive_id]