+++ spec/models/loader/modify_policy_spec.rb
@@ -131,19 +131,6 @@ RSpec.describe(Loader::ModifyPolicy) do
     end
   end
 
-  let(:diff_response) do
-    ::SuccessResponse.new(
-      DB::Repository::DataObjects::DiffElements.new(
-        annotations: annotations,
-        credentials: credentials,
-        permissions: permissions,
-        resources: resources,
-        role_memberships: role_memberships,
-        roles: roles
-      )
-    )
-  end
-
   let(:policy_repository) do
     DB::Repository::PolicyRepository.new(db: db).tap do |repo|
       allow(repo).to receive(:find_created_elements)
@@ -158,8 +145,31 @@ RSpec.describe(Loader::ModifyPolicy) do
     end
   end
 
+  let(:diff_response) do
+    DB::Repository::DataObjects::DiffElements.new(
+      annotations: annotations,
+      credentials: credentials,
+      permissions: permissions,
+      resources: resources,
+      role_memberships: role_memberships,
+      roles: roles
+    )
+  end
+
+  let(:policy_diff_response) do
+    ::SuccessResponse.new({
+      created: diff_response,
+      deleted: diff_response,
+      updated: diff_response,
+      final: diff_response
+    })
+  end
+
   let(:policy_diff) do
-    CommandHandler::PolicyDiff.new(policy_repository: policy_repository)
+    CommandHandler::PolicyDiff.new(policy_repository: policy_repository).tap do |mock|
+      allow(mock).to receive(:call)
+        .and_return(policy_diff_response)
+    end
   end
 
   describe '.call' do
@@ -193,7 +203,7 @@ RSpec.describe(Loader::ModifyPolicy) do
           let(:role_memberships) { [] }
           let(:roles) { [] }
 
-          it 'the diff diff result does not include annotations' do
+          it 'the diff result includes the annotations' do
             result = subject.call
 
             # The PoliyResult is valid
@@ -242,7 +252,7 @@ RSpec.describe(Loader::ModifyPolicy) do
           let(:role_memberships) { [] }
           let(:roles) { [] }
 
-          it 'the diff result does not include credentials' do
+          it 'the diff result includes the credentials' do
             result = subject.call
 
             # The PoliyResult is valid
@@ -291,7 +301,7 @@ RSpec.describe(Loader::ModifyPolicy) do
           let(:role_memberships) { [] }
           let(:roles) { [] }
 
-          it 'the diff result does not include permissions' do
+          it 'the diff result includes the permissions' do
             result = subject.call
 
             # The PoliyResult is valid
@@ -340,7 +350,7 @@ RSpec.describe(Loader::ModifyPolicy) do
           let(:role_memberships) { [] }
           let(:roles) { [] }
 
-          it 'the diff result does not include resources' do
+          it 'the diff result includes the resources' do
             result = subject.call
 
             # The PoliyResult is valid
@@ -389,7 +399,7 @@ RSpec.describe(Loader::ModifyPolicy) do
           let(:resources) { [] }
           let(:roles) { [] }
 
-          it 'the diff result does not include role memberships' do
+          it 'the diff result includes the role memberships' do
             result = subject.call
 
             # The PoliyResult is valid
@@ -438,7 +448,7 @@ RSpec.describe(Loader::ModifyPolicy) do
           let(:resources) { [] }
           let(:role_memberships) { [] }
 
-          it 'the diff result does not include roles' do
+          it 'the diff result includes the roles' do
             result = subject.call
 
             # The PoliyResult is valid
@@ -496,30 +506,30 @@ RSpec.describe(Loader::ModifyPolicy) do
 
             expect(result.diff[:created])
               .to be_a(DB::Repository::DataObjects::DiffElements)
-            expect(result.diff[:created].annotations.length).to eq(1)  
-            expect(result.diff[:created].credentials.length).to eq(1)  
-            expect(result.diff[:created].permissions.length).to eq(1)  
-            expect(result.diff[:created].resources.length).to eq(1)  
-            expect(result.diff[:created].role_memberships.length).to eq(1)  
-            expect(result.diff[:created].roles.length).to eq(1)  
+            expect(result.diff[:created].annotations.length).to eq(1)
+            expect(result.diff[:created].credentials.length).to eq(1)
+            expect(result.diff[:created].permissions.length).to eq(1)
+            expect(result.diff[:created].resources.length).to eq(1)
+            expect(result.diff[:created].role_memberships.length).to eq(1)
+            expect(result.diff[:created].roles.length).to eq(1)
 
             expect(result.diff[:deleted])
               .to be_a(DB::Repository::DataObjects::DiffElements)
-            expect(result.diff[:deleted].annotations.length).to eq(1)  
-            expect(result.diff[:deleted].credentials.length).to eq(1)  
-            expect(result.diff[:deleted].permissions.length).to eq(1)  
-            expect(result.diff[:deleted].resources.length).to eq(1)  
-            expect(result.diff[:deleted].role_memberships.length).to eq(1)  
-            expect(result.diff[:deleted].roles.length).to eq(1)  
+            expect(result.diff[:deleted].annotations.length).to eq(1)
+            expect(result.diff[:deleted].credentials.length).to eq(1)
+            expect(result.diff[:deleted].permissions.length).to eq(1)
+            expect(result.diff[:deleted].resources.length).to eq(1)
+            expect(result.diff[:deleted].role_memberships.length).to eq(1)
+            expect(result.diff[:deleted].roles.length).to eq(1)
 
             expect(result.diff[:updated])
               .to be_a(DB::Repository::DataObjects::DiffElements)
-            expect(result.diff[:updated].annotations.length).to eq(1)  
-            expect(result.diff[:updated].credentials.length).to eq(1)  
-            expect(result.diff[:updated].permissions.length).to eq(1)  
-            expect(result.diff[:updated].resources.length).to eq(1)  
-            expect(result.diff[:updated].role_memberships.length).to eq(1)  
-            expect(result.diff[:updated].roles.length).to eq(1)  
+            expect(result.diff[:updated].annotations.length).to eq(1)
+            expect(result.diff[:updated].credentials.length).to eq(1)
+            expect(result.diff[:updated].permissions.length).to eq(1)
+            expect(result.diff[:updated].resources.length).to eq(1)
+            expect(result.diff[:updated].role_memberships.length).to eq(1)
+            expect(result.diff[:updated].roles.length).to eq(1)
           end
         end
       end