+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `dryRun` parameter, to report policy attributes that would be created by the 
   submitted policy.
   CNJR-6999
+- Modifies the REST API response of a Policy load command, when called with the
+  `dryRun` parameter, to report policy attributes that would be updated by the 
+  submitted policy.
+  CNJR-6109
 
 ### Fixed
 - Updates audit events generated during Policy Factory usage.